탈퇴 완료 및 로그아웃

+++ src/main/java/site/metacoding/miniproject/service/CompanyService.java
@@ -69,8 +69,14 @@ public class CompanyService {
   }
 
   public void deleteCompany(Integer companyId) {
-    companyDao.deleteById(companyId);
-    coCheckDao.deleteById(companyId);
+    Company companyPS = companyDao.findById(companyId);
+    if (companyPS != null) {
+      companyDao.deleteById(companyId);
+      coCheckDao.deleteById(companyId);
+    } else {
+      throw new RuntimeException("해당 " + companyId + "로 삭제를 할 수 없습니다.");
+    }
+
   }
 
   public boolean usernameSameCheck(String companyUsername) {
